+++ packages/fury/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Fury Changelog
 
-## Master
+## 3.0.0-beta.11 (2019-06-11)
 
 ### Breaking